enhance and clean code
enhance and clean code and separate logic
