+++ Currancy exchange/feature/tabviews/Favourite/FavouriteView.swift
@@ -1,0 +1,22 @@
+//
+//  FavouriteView.swift
+//  Currancy exchange
+//
+//  Created by Tayseer-Anwar on 13/10/2024.
+//
+
+import SwiftUI
+
+struct FavouriteView: View {
+    var body: some View {
+        VStack{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+        .frame(maxWidth: .infinity,maxHeight: .infinity)
+        .background(.yellow)
+    }
+}
+
+#Preview {
+    FavouriteView()
+}
