+++ Currancy exchange/feature/tabviews/Market/MarketView.swift
@@ -1,0 +1,22 @@
+//
+//  MarketView.swift
+//  Currancy exchange
+//
+//  Created by Tayseer-Anwar on 13/10/2024.
+//
+
+import SwiftUI
+
+struct MarketView: View {
+    var body: some View {
+        VStack{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+        .frame(maxWidth: .infinity,maxHeight: .infinity)
+        .background(.pink)
+    }
+}
+
+#Preview {
+    MarketView()
+}
