+++ Currancy exchange/feature/tabviews/Tab/ViewModel/TabViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  TabViewModel.swift
+//  Currancy exchange
+//
+//  Created by Tayseer-Anwar on 13/10/2024.
+//
+
+import Foundation
+
+
+class TabViewModel: ObservableObject { // Generic struct 'StateObject' requires that 'TabViewModel' conform to 'ObservableObject'
+    @Published var currentView: Tab = .Home
+    let tabs: [Tab] = [.Home, .Trade, .Market, .Favourite, .Wallet]
+    
+    
+}
