+++ Currancy exchange/feature/tabviews/Trade/TradeView.swift
@@ -1,0 +1,22 @@
+//
+//  TradeView.swift
+//  Currancy exchange
+//
+//  Created by Tayseer-Anwar on 13/10/2024.
+//
+
+import SwiftUI
+
+struct TradeView: View {
+    var body: some View {
+        VStack{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+        .frame(maxWidth: .infinity,maxHeight: .infinity)
+        .background(.red)
+    }
+}
+
+#Preview {
+    TradeView()
+}
