+++ Currancy exchange/feature/tabviews/Wallet/WalletView.swift
@@ -1,0 +1,22 @@
+//
+//  WalletView.swift
+//  Currancy exchange
+//
+//  Created by Tayseer-Anwar on 13/10/2024.
+//
+
+import SwiftUI
+
+struct WalletView: View {
+    var body: some View {
+        VStack{
+            Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        }
+        .frame(maxWidth: .infinity,maxHeight: .infinity)
+        .background(.purple)
+    }
+}
+
+#Preview {
+    WalletView()
+}
